pyproject.toml
[tool.poetry]
name = "matching-data"
version = "0.1.0"
description = ""
authors = ["jose9806"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
lxml = "^5.3.1"
rapidfuzz = "^3.12.2"
tqdm = "^4.67.1"
psutil = "^7.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
